allow adding third party protobuf
We want to use the `google.protobuf.FieldMask` for updating shares. For that we need to be able to use third party protobuf ... or we switch to ̀[`buf`](https://github.com/bufbuild/buf), since `prototool` is in maintenance mode only and actually recommends `buf` in their readme.

go glob does not support **

Signed-off-by: Jörn Friedrich Dreyer <jfd@butonic.de>